Reorganized constants and made preprocessing file
