+++ assignments.py
@@ -2,6 +2,7 @@
 Includes group optimization for common pickup locations.
 """
 
+from constants import *
 from datetime import date
 import numpy as np
 import pandas as pd
