+++ constants.py
@@ -1,0 +1,32 @@
+
+OUTPUT_DRIVER_NAME_KEY = 'Driver'
+OUTPUT_DRIVER_PHONE_KEY = 'Driver Phone #'
+
+DRIVER_TIMESTAMP_KEY = 'Timestamp'
+DRIVER_NAME_KEY = 'Name'
+DRIVER_PHONE_KEY = 'Phone Number'
+DRIVER_CAPACITY_KEY = 'Number of Seats in Car (not including you)'
+
+RIDER_TIMESTAMP_KEY = 'Timestamp'
+RIDER_NAME_KEY = 'Rider'
+RIDER_PHONE_KEY = 'Rider Phone #'
+RIDER_LOCATION_KEY = 'Location'
+RIDER_FRIDAY_KEY = 'Friday'
+RIDER_SUNDAY_KEY = 'Sunday'
+RIDER_NOTES_KEY = 'Notes'
+
+PERMANENT_RIDER_TIMESTAMP_KEY = 'Timestamp'
+PERMANENT_RIDER_NAME_KEY = 'Full Name:'
+PERMANENT_RIDER_PHONE_KEY = 'Phone Number: '
+PERMANENT_RIDER_LOCATION_KEY = 'Where should we pick you up?'
+PERMANENT_RIDER_FRIDAY_KEY = 'Which service(s) do you need a permanent ride for? [Friday Night Bible Study | 6:30 pm]'
+PERMANENT_RIDER_SUNDAY_KEY = 'Which service(s) do you need a permanent ride for? [Sunday Service | 8:30 am/10:45 am]'
+PERMANENT_RIDER_NOTES_KEY = 'Other Notes'
+
+WEEKLY_RIDER_TIMESTAMP_KEY = 'Timestamp'
+WEEKLY_RIDER_NAME_KEY = 'Name (Include your last name if this is your first time)'
+WEEKLY_RIDER_PHONE_KEY = 'Phone Number '
+WEEKLY_RIDER_LOCATION_KEY = 'Where should we pick you up from?'
+WEEKLY_RIDER_FRIDAY_KEY = 'Friday Night Bible Study (Friday @7pm) (Rides from Campus will be provided at Peterson Loop at 6:30 pm)'
+WEEKLY_RIDER_SUNDAY_KEY = 'Sunday Service '
+WEEKLY_RIDER_NOTES_KEY = 'Additional Comments / Questions / Concerns'
